chore: get tests running through karma with system.js
 - TODO: entity serialization tests are failing and need the System.js loader code promoted from angular2-rpg.
