+++ karma.conf.js
@@ -8,11 +8,14 @@ module.exports = function (config) {
     basePath: '',
     frameworks: ['jasmine'],
     files: [
-      "test/vendor/jquery/dist/jquery.min.js",
-      "test/vendor/underscore/underscore-min.js",
+      "vendor/jquery/dist/jquery.min.js",
+      "vendor/underscore/underscore-min.js",
+      'node_modules/es6-module-loader/dist/es6-module-loader-sans-promises.src.js',
+      'node_modules/systemjs/dist/system.src.js',
       "lib/pow-core.js",
-      "lib/test/*.js",
-      {pattern: 'test/fixtures/*.*', watched: true, included: false, served: true}
+      {pattern: 'test/fixtures/*.*', watched: true, included: false, served: true},
+      {pattern: 'test/**/*.js', watched: true, included: false, served: true},
+      "karma.main.js"
     ],
     reporters: ['dots', 'coverage'],
     port: 9876,
